refactor: move multica registry to surface package
Move Multica participant registry state into the surface/multica adapter package while keeping driver facade functions for existing command and runtime callers.

Validation: go test ./harness/internal/surface/multica ./harness/internal/driver ./harness/cmd/mnemon-harness ./harness/cmd/mnemon-multica-runtime ./harness/cmd/mnemon-acceptance
